refactor: rendering logic
