+++ src/ui/chrome.rs
@@ -1,0 +1,112 @@
+use std::io::Write;
+
+use crossterm::{cursor, queue, style::Color};
+
+use crate::app_state::AppState;
+use crate::ui::text::{display_width, print_colored_text, truncate_to_width};
+
+pub fn print_loading_indicator<W: Write>(stdout: &mut W, cols: u16, rows: u16) {
+    // Center the loading message
+    let message = "Loading...";
+    let x = (cols.saturating_sub(message.len() as u16)) / 2;
+    let y = rows / 2;
+
+    queue!(stdout, cursor::MoveTo(x, y)).unwrap();
+    print_colored_text(stdout, message, Color::Yellow, None, None);
+}
+
+pub fn print_function_keys<W: Write>(
+    stdout: &mut W,
+    cols: u16,
+    rows: u16,
+    state: &AppState,
+    is_remote: bool,
+) {
+    // Move to bottom of screen
+    queue!(stdout, cursor::MoveTo(0, rows - 1)).unwrap();
+
+    // Get current sorting indicator
+    let sort_indicator = match state.sort_criteria {
+        crate::app_state::SortCriteria::Default => "Sort:Default",
+        crate::app_state::SortCriteria::Pid => "Sort:PID",
+        crate::app_state::SortCriteria::Memory => "Sort:Memory",
+        crate::app_state::SortCriteria::Utilization => "Sort:Util",
+        crate::app_state::SortCriteria::GpuMemory => "Sort:GPU-Mem",
+        crate::app_state::SortCriteria::Power => "Sort:Power",
+        crate::app_state::SortCriteria::Temperature => "Sort:Temp",
+    };
+
+    let function_keys = if is_remote {
+        // Remote mode: only GPU sorting
+        format!(
+            "h:Help q:Exit ←→:Tabs ↑↓:Scroll PgUp/PgDn:Page d:Default u:Util g:GPU-Mem [{sort_indicator}]"
+        )
+    } else {
+        // Local mode: both process and GPU sorting
+        format!("h:Help q:Exit ←→:Tabs ↑↓:Scroll PgUp/PgDn:Page p:PID m:Memory d:Default u:Util g:GPU-Mem [{sort_indicator}]")
+    };
+
+    let truncated_keys = if display_width(&function_keys) > cols as usize {
+        truncate_to_width(&function_keys, cols as usize)
+    } else {
+        function_keys
+    };
+
+    // Check if there's a notification to display
+    let notification_msg = state.notifications.get_current_message().unwrap_or("");
+    let notification_len = display_width(notification_msg);
+
+    // Calculate space available for function keys (reserve space for notification)
+    let available_space = if notification_len > 0 {
+        cols.saturating_sub(notification_len as u16 + 1) // +1 for separator space
+    } else {
+        cols
+    } as usize;
+
+    // Truncate function keys if necessary to make room for notification
+    let final_function_keys = if display_width(&truncated_keys) > available_space {
+        truncate_to_width(&truncated_keys, available_space)
+    } else {
+        truncated_keys
+    };
+
+    // Print function keys
+    print_colored_text(stdout, &final_function_keys, Color::DarkGreen, None, None);
+
+    // Print notification if there is one
+    if notification_len > 0 {
+        // Add separator
+        print_colored_text(stdout, " ", Color::White, None, None);
+
+        // Print notification with appropriate color
+        let notification_color =
+            if notification_msg.contains("Error") || notification_msg.contains("Failed") {
+                Color::Red
+            } else if notification_msg.contains("Warning") {
+                Color::Yellow
+            } else {
+                Color::Cyan
+            };
+
+        print_colored_text(stdout, notification_msg, notification_color, None, None);
+    }
+
+    // Fill remaining space to clear any leftover text
+    let used_space = display_width(&final_function_keys)
+        + if notification_len > 0 {
+            notification_len + 1
+        } else {
+            0
+        };
+    let remaining_space = cols as usize - used_space.min(cols as usize);
+
+    if remaining_space > 0 {
+        print_colored_text(
+            stdout,
+            &" ".repeat(remaining_space),
+            Color::White,
+            None,
+            None,
+        );
+    }
+}
